Move intermediate content into new notebook
The combined content includes the pandas
categorical, string, profiling, and timing content
from my MDS seminar,
which was merged from the int-pand-jl remote branch
in the previous commit.
It also includes the notebook with EDA tips
that I created for MDS separately
and added in this commit.
Together,
these will constitute most of the extracurricular material.